+++ mbgdml/data/__init__.py
@@ -1,7 +1,6 @@
 """Data structures handled by mbgdml."""
 
 from .basedata import mbGDMLData
-from .model import mbModel
 from .predictset import predictSet
 from .dataset import dataSet